requirements.txt
# XParking - Smart Parking System Dependencies
# Install: pip install -r requirements.txt

# === Core ===
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# === GUI ===
# tkinter is built-in with Python, no need to install

# === MQTT Communication ===
paho-mqtt>=1.6.0

# === HTTP Requests ===
requests>=2.31.0

# === HTTP Server (for ESP32-CAM) ===
flask>=3.0.0

# === AI/ML - License Plate Recognition ===
torch>=2.0.0
torchvision>=0.15.0

# === QR Code ===
qrcode>=7.4.0
pyzbar>=0.1.9

# === OCR (Optional - for Tesseract fallback) ===
pytesseract>=0.3.10

# === Email ===
# smtplib is built-in with Python, no need to install

# === YOLOv5 dependencies ===
matplotlib>=3.7.0
scipy>=1.10.0
tqdm>=4.65.0
seaborn>=0.12.0
pandas>=2.0.0
PyYAML>=6.0
gitpython>=3.1.0
psutil>=5.9.0
